fix: TransitView 배경 적용이 안되던 버그 수정
